+++ src/collective/chimpfeed/form.py
@@ -1,7 +1,9 @@
 import time
+import datetime
 import math
 import greatape
 import cgi
+import urllib
 
 from Products.statusmessages.interfaces import IStatusMessage
 from Products.CMFCore.utils import getToolByName
@@ -34,6 +36,7 @@ from collective.chimpfeed import MessageFactory as _
 from collective.chimpfeed import logger
 from collective.chimpfeed.interfaces import IFeedSettings
 from collective.chimpfeed.interfaces import INameSplitter
+from collective.chimpfeed.interfaces import ICampaignPortlet
 from collective.chimpfeed.vocabularies import interest_groups_factory
 from collective.chimpfeed.splitters import GenericNameSplitter
 
@@ -50,6 +53,18 @@ def cache_on_get_for_an_hour(method, self):
         raise DontCache
 
     return time.time() // (60 * 60), self.id
+
+
+class ICampaign(ICampaignPortlet):
+    """Note that most fields are inherited from the portlet."""
+
+    schedule = schema.Datetime(
+        title=_(u"Schedule date"),
+        description=_(u"If provided, item will be scheduled to be sent "
+                      u"at this time."),
+        required=False,
+        default=None,
+        )
 
 
 class IModeration(Interface):
@@ -196,7 +211,7 @@ class InterestsWidget(SequenceWidget):
 
         for grouping in filtered:
             terms = tuple(interest_groups_factory.get_terms_for_grouping(
-                grouping, count=False
+                grouping
                 ))
 
             # Create label from the grouping name; we simply lowercase
@@ -262,6 +277,179 @@ class BaseForm(form.Form):
             u64(self.context._p_oid),
             int(self.context._p_mtime) % 10000
             )
+
+
+class CampaignForm(BaseForm):
+    fields = field.Fields(
+        ICampaign['start'],
+        ICampaign['subject'],
+        ICampaign['schedule'],
+        )
+
+    ignoreContext = True
+
+    @button.buttonAndHandler(_(u'Preview'))
+    def handlePreview(self, action):
+        data, errors = self.extractData()
+        if errors:
+            self.status = self.formErrorsMessage
+            return
+
+        url = self.context.portal_url()
+
+        return self.request.response.redirect(
+            url + "/@@chimpfeed-preview?%s" % urllib.urlencode({
+                'start': data['start'].isoformat(),
+                'image': self.context.image,
+                'scale': self.context.scale
+                }))
+
+    @button.buttonAndHandler(_(u'Create'))
+    def handleCreate(self, action):
+        data, errors = self.extractData()
+        if errors:
+            self.status = self.formErrorsMessage
+            return
+
+        success = self.process("campaignCreate", **data)
+
+        if success:
+            IStatusMessage(self.request).addStatusMessage(
+                _(u"Campaign created."),
+                "info"
+                )
+
+            # Set date to tomorrow's date.
+            self.context.start = datetime.date.today() + \
+                                 datetime.timedelta(days=1)
+
+    def process(self, method, subject=None, start=None, schedule=None):
+        settings = IFeedSettings(self.context)
+        api_key = settings.mailchimp_api_key
+
+        site = self.context.portal_url.getPortalObject()
+        view = site.restrictedTraverse('chimpfeed-campaign')
+
+        rendered = view.template(
+            start=start.isoformat(),
+            image=self.context.image,
+            scale=self.context.scale,
+            ).encode('utf-8')
+
+        next_url = self.request.get('HTTP_REFERER') or self.action
+
+        try:
+            if not rendered.strip():
+                IStatusMessage(self.request).addStatusMessage(
+                    _(u"No content found; newsletter not created."),
+                    "info"
+                    )
+
+                return
+
+            if api_key:
+                api = greatape.MailChimp(api_key, debug=False)
+
+                try:
+                    for entry in api(method="lists"):
+                        if entry['id'] == self.context.mailinglist:
+                            break
+                    else:
+                        IStatusMessage(self.request).addStatusMessage(
+                            _(u"Mailinglist not found."),
+                            "error"
+                            )
+
+                        return
+
+                    if self.context.template:
+                        section = 'html_%s' % self.context.section
+                    else:
+                        section = 'html'
+
+                    result = api(
+                        method=method,
+                        type="regular",
+                        options={
+                            'subject': subject or entry['default_subject'],
+                            'from_email': entry['default_from_email'],
+                            'from_name': entry['default_from_name'],
+                            'to_email': 0,
+                            'list_id': self.context.mailinglist,
+                            'template_id': self.context.template or None,
+                            'generate_text': True,
+                            },
+                        content={
+                            section: rendered,
+                            },
+                        )
+
+                    if result:
+                        if schedule:
+                            # Apply local time zone to get GMT
+                            schedule = schedule + datetime.timedelta(
+                                seconds=time.timezone
+                                )
+
+                            # Format time
+                            schedule_time = time.strftime(
+                                "%Y-%m-%d %H:%M:%S", schedule.utctimetuple()
+                                )
+
+                            schedule = api(
+                                method="campaignSchedule",
+                                cid=result,
+                                schedule_time=schedule_time,
+                                )
+
+                            if not schedule:
+                                IStatusMessage(self.request).addStatusMessage(
+                                    _(u"Campaign created, but not scheduled."),
+                                    "error"
+                                    )
+
+                                return
+
+                        next_url = self.context.portal_url() + \
+                                   "/@@chimpfeed-content?cid=%s" % result
+
+                except greatape.MailChimpError, exc:
+                    IStatusMessage(self.request).addStatusMessage(
+                        _(u"Unable to process request."), "error"
+                        )
+                    logger.warn(exc.msg)
+                    return
+
+                return bool(result)
+        finally:
+            self.request.response.redirect(next_url)
+
+    def update(self):
+        super(CampaignForm, self).update()
+
+        today = datetime.date.today()
+
+        date = self.context.start
+        if date is None:
+            date = today
+
+        start = self.widgets['start']
+        start.value = (date.year, date.month, date.day)
+
+        subject = self.widgets['subject']
+        if not subject.value:
+            subject.value = self.context.subject or \
+                            self.context.Title().decode('utf-8')
+
+        schedule = self.widgets['schedule']
+        assert isinstance(schedule.value, tuple)
+
+        if not schedule.value[0]:
+            tomorrow = today + datetime.timedelta(days=1)
+
+            schedule.value = (
+                tomorrow.year, tomorrow.month, tomorrow.day, "09", "00"
+                )
 
 
 class ModerationForm(BaseForm):