+++ src/collective/chimpfeed/portlet.py
@@ -28,19 +28,38 @@ from zope.security import checkPermission
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.CMFDefault.interfaces import ICMFDefaultSkin
 
+from collective.chimpfeed.form import CampaignForm
 from collective.chimpfeed.form import SubscribeForm
 from collective.chimpfeed.form import ModerationForm
 from collective.chimpfeed.interfaces import ISubscriptionPortlet
 from collective.chimpfeed.interfaces import IModerationPortlet
+from collective.chimpfeed.interfaces import ICampaignPortlet
 from collective.chimpfeed import MessageFactory as _
+
+
+class CampaignPortletAssignment(base.Assignment):
+    implements(ICampaignPortlet)
+
+    heading = _(u"Campaign scheduling")
+    description = _(u"Send or schedule a newsletter campaign.")
+
+    title = _(u"Campaign portlet")
+
+    start = None
+    section = u"std_content00"
+
+    def __init__(self, **kwargs):
+        for name, value in kwargs.items():
+            self.__dict__[name] = value
 
 
 class ModerationPortletAssignment(base.Assignment):
     implements(IModerationPortlet)
 
-    heading = _(u"Moderation")
+    heading = _(u"Feed moderation")
+    description = _(u"Use this form to moderate content.")
 
-    title = _(u"Feed moderation")
+    title = _(u"Moderation portlet")
 
 
 class SubscriptionPortletAssignment(base.Assignment):
@@ -97,6 +116,29 @@ class SubscriptionPortletRenderer(FormPortletRenderer):
     def create_form(self):
         context = self.data.__of__(self.context)
         return SubscribeForm(context, self.request)
+
+
+class CampaignPortletRenderer(FormPortletRenderer):
+    def create_form(self):
+        context = self.data.__of__(self.context)
+        return CampaignForm(context, self.request)
+
+
+class CampaignPortletAddForm(base.AddForm):
+    label = _(u"Add campaign portlet")
+    form_fields = form.Fields(ICampaignPortlet)
+
+    # Let's not ask the user to name a start date on adding the
+    # portlet.
+    form_fields = form_fields.omit('start')
+
+    def create(self, data):
+        return CampaignPortletAssignment(**data)
+
+
+class CampaignPortletEditForm(base.EditForm):
+    label = _(u"Edit campaign portlet")
+    form_fields = form.Fields(ICampaignPortlet)
 
 
 class ModerationPortletAddForm(base.NullAddForm):